# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Tegan Straley Web API assignment 2
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /deletes:
    # binds swagger app logic to a route
    x-swagger-router-controller: Node
    delete:
      description: deletes for delete
      # used as the method name of the controller
      operationId: deleteNode
      parameters:
        - name: Body
          in: query
          description: Key/value pairs
          required: false
          type: string
      responses:
        "200":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string